+++ Core/Models/Account/MerchantRegisterDto.cs
@@ -24,6 +24,6 @@ namespace Refundeo.Core.Models.Account
         public string ContactPhone { get; set; }
         public string Currency { get; set; }
         public IList<OpeningHoursDto> OpeningHours { get; set; }
-        public IList<long> Tags { get; set; }
+        public IList<int> Tags { get; set; }
     }
 }
